requirements.txt
numpy
scipy
matplotlib
opencv
pytesseract
beautifulsoup4
lxml
fastapi
snapshottest
graphene
ipython
aiohttp
pydantic
pytest-asyncio
uvicorn
